+++ test/browser/popupElement.test.jsx
@@ -93,7 +93,7 @@ describe("popupElement", () => {
     await page.hover("input");
     let t = await page.evaluate(() => ({ ...t, html: testEl.outerHTML, isOpened: testEl.$isOpen }));
     expect(t.isOpened).toBe(false); // because of timeout
-    await page.waitForTimeout(220); // hoverShowTimeout is 200 by default
+    await page.waitForTimeout(300); // hoverShowTimeout is 200 by default
     t = await page.evaluate(() => ({ ...t, html: testEl.outerHTML, isOpened: testEl.$isOpen }));
     expect(t.isOpened).toBe(true);